Apply the replacement of usages for all the ocladdressnodes associated with pi

+++ tornado-runtime/src/main/java/uk/ac/manchester/tornado/runtime/graal/phases/TornadoNativeTypeElimination.java
@@ -115,12 +115,15 @@ public class TornadoNativeTypeElimination extends BasePhase<TornadoSketchTierCon
         // identify the usages that need to be removed
         for (Node node : fixedGuardNode.usages()) {
             if (node instanceof PiNode pi && (pi.usages().filter(OffsetAddressNode.class).isNotEmpty())) {
-                OffsetAddressNode off = pi.usages().filter(OffsetAddressNode.class).first();
-                // if this address node is used by a javaread/javawrite node
-                if (off.usages().filter(JavaReadNode.class).isNotEmpty() //
-                        || off.usages().filter(JavaWriteNode.class).isNotEmpty() //
-                        || off.usages().filter(WriteAtomicNode.class).isNotEmpty()) {
-                    off.replaceFirstInput(pi, loadFieldSegment);
+                // System.out.println(">> Remove pi " + pi);
+                for (OffsetAddressNode off : pi.usages().filter(OffsetAddressNode.class)) {
+                    //OffsetAddressNode off = pi.usages().filter(OffsetAddressNode.class).first();
+                    // if this address node is used by a javaread/javawrite node
+                    if (off.usages().filter(JavaReadNode.class).isNotEmpty() //
+                            || off.usages().filter(JavaWriteNode.class).isNotEmpty() //
+                            || off.usages().filter(WriteAtomicNode.class).isNotEmpty()) {
+                        off.replaceFirstInput(pi, loadFieldSegment);
+                    }
                     nodesToBeRemoved.add(pi);
                 }
             } else if (node instanceof PiNode pi && (pi.usages().filter(LoadHubNode.class).isNotEmpty())) {
@@ -156,7 +159,7 @@ public class TornadoNativeTypeElimination extends BasePhase<TornadoSketchTierCon
 
         for (int i = 0; i < nodesToBeRemoved.size(); i++) {
             Node n = nodesToBeRemoved.get(i);
-            if (n != null) {
+            if (n != null && !n.isDeleted()) {
                 if (n instanceof FixedGuardNode) {
                     deleteFixed(n);
                 } else {
